ALMOST working; unexpected rapid motion after transition to state_inserting_peg caused a major collision. I dont know how to track this down, so now may be the time to move permanently to a hand-coded motion system using only the cartesian_force_controller, leaving compliance_controller behind for good. Its just too dangerous.
